Move bracket handling to the proper classes
